Add bin/loop as hack for rpi bootup issues
To run as an independent gizmo, the raspberry pi zero needs to run the script on boot.
I investigated two approaches, each with issues:
- systemd unit
- kick off script for login shell in .bashrc

The systemd unit approach had rad pros (gets logging, retries, waiting
for network for free), but a fatal flaw (it runs outside of tty1, so it
doesn't get the keyboard input events).  The .bashrc approach had weaker
conceptual pros (delightfully simple hack), but, critically, it actually
works.

One important note is that the actual script invocation in the .bashrc
must be inside a conditional guard like the following:
```sh
if [[ $(tty) == /dev/tty1 ]]; then
  # do ya thing
fi
```
or it will hijack ssh sessions too. Use a sharp axe when hacking.